+++ src/hooks/usePendingActions.ts
@@ -22,16 +22,19 @@ const usePendingActions = (chainId: string, safeAddress?: string): PendingAction
   const { safeAddress: currentSafeAddress } = useSafeInfo()
   const { page: currentSafeQueue } = useTxQueue()
   const isCurrentSafe = currentSafeAddress === safeAddress
-  const cachedQueue = useRef<Promise<TransactionListPage>>()
+  const cachedQueue = useRef<TransactionListPage>()
 
-  const [loadedQueue] = useAsync<TransactionListPage>(() => {
+  const [loadedQueue] = useAsync<TransactionListPage | undefined>(async () => {
     if (isCurrentSafe || !safeAddress) return
 
-    if (!cachedQueue.current) {
-      cachedQueue.current = getTransactionQueue(chainId, safeAddress)
+    if (cachedQueue.current) {
+      return cachedQueue.current
     }
 
-    return cachedQueue.current
+    const fetchedQueue = await getTransactionQueue(chainId, safeAddress)
+    cachedQueue.current = fetchedQueue
+
+    return fetchedQueue
   }, [chainId, safeAddress, isCurrentSafe])
 
   const queue = isCurrentSafe ? currentSafeQueue : loadedQueue
